코드 리팩토링
코드 리팩토링 및 date.ts 위치 변경
