--- src/components/map/initMap.tsx
@@ -1,11 +1,0 @@
-"use client"
-import {GoogleMapWrapper} from "@/components/map/mapComponents/GoogleMapWrapper";
-import {GoogleMap} from "@/components/map/mapComponents/GoogleMap";
-
-export const MapComponent = () => {
-    return(
-        <GoogleMapWrapper>
-            <GoogleMap/>
-        </GoogleMapWrapper>
-    )
-}
